requirements.txt
ag-ui-protocol==0.1.13
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
appdirs==1.4.4
APScheduler==3.11.2
argcomplete==3.6.3
attrs==25.4.0
Authlib==1.6.8
azure-core==1.38.2
azure-identity==1.25.2
azure-storage-blob==12.28.0
backoff==2.2.1
bcrypt==5.0.0
beartype==0.22.9
beautifulsoup4==4.13.5
boto3==1.40.76
botocore==1.40.76
branca==0.8.2
build==1.4.0
cachetools==7.0.1
caio==0.9.25
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
cohere==5.20.6
crewai==1.9.3
crewai-tools==1.9.3
croniter==6.0.0
cross-web==0.4.1
cryptography==46.0.5
cyclopts==4.5.4
dataclasses-json==0.6.7
decorator==5.2.1
deepeval==3.8.6
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
execnet==2.1.2
executing==2.2.1
fastapi==0.129.2
fastapi-sso==0.16.0
fastavro==1.12.1
fastmcp==3.0.2
fastuuid==0.14.0
filelock==3.24.3
flatbuffers==25.12.19
folium==0.20.0
frozenlist==1.8.0
fsspec==2026.2.0
genai-prices==0.0.54
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
graphql-core==3.2.7
griffelib==2.0.0
groq==1.0.0
grpcio==1.78.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
hyperframe==6.1.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.12.0
invoke==2.2.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.1.0
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.1
jsonschema-specifications==2025.9.1
keyring==25.7.0
kubernetes==35.0.0
lancedb==0.5.7
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.8
langsmith==0.7.6
litellm==1.81.14
litellm-enterprise==0.1.32
litellm-proxy-extras==0.4.46
logfire==4.25.0
logfire-api==4.25.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.23.3
mdurl==0.1.2
mistralai==1.12.4
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nexus-rpc==1.2.0
nodeenv==1.10.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openai==1.83.0
openapi-pydantic==0.5.1
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pathable==0.5.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
py==1.11.0
py-key-value-aio==0.4.4
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-ai==1.62.0
pydantic-ai-slim==1.62.0
pydantic-evals==1.62.0
pydantic-graph==1.62.0
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.11.0
pylance==0.9.18
PyMuPDF==1.26.7
PyNaCl==1.6.2
pyparsing==3.3.2
pypdfium2==5.5.0
pyperclip==1.11.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyroscope-io==0.8.16
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.22
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.17.0
ratelimiter==1.2.0.post0
redis==7.2.0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rich-rst==1.3.2
rpds-py==0.30.0
rq==2.7.0
rsa==4.9.1
s3transfer==0.14.0
semver==3.0.4
sentry-sdk==2.53.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.52.1
strawberry-graphql==0.307.1
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.21
temporalio==1.20.0
tenacity==9.1.4
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.67.3
typer==0.23.1
typer-slim==0.24.0
types-protobuf==6.32.1.20260221
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.9.30
uvicorn==0.41.0
uvloop==0.21.0
virtualenv==20.38.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==15.0.1
wheel==0.46.3
wrapt==1.17.3
xai-sdk==1.7.0
xxhash==3.6.0
xyzservices==2025.11.0
yarl==1.22.0
youtube-transcript-api==1.2.4
zipp==3.23.0
zstandard==0.25.0
